updated main and homepage hndlbrs and udpated parent-route typo
