+++ controllers/home-routes.js
@@ -1,0 +1,7 @@
+const router = require('express').Router();
+
+router.get('/', (req, res) => {
+  res.render('homepage');
+});
+
+module.exports = router;
