+++ controllers/index.js
@@ -1,8 +1,10 @@
 const router = require('express').Router();
+const homeRoutes = require('./home-routes.js');
 
 const apiRoutes = require('./api');
 
 router.use('/api', apiRoutes);
+router.use('/', homeRoutes);
 
 router.use((req, res) => {
   res.status(404).end();
