+++ public/javascript/login.js
@@ -1,33 +1,5 @@
 // functions for main page - sign up and/or login? Confirm/review with team  All css data just place holders until receive from Peter
 
-async function signupFormHandler(event) {
-  event.preventDefault();
-
-  const parent_name = document.querySelector('#parent-name-signup').value.trim();
-  const phone = document.querySelector('#phone-signup').value.trim();
-  const email = document.querySelector('#email-signup').value.trim();
-  const password = document.querySelector('#password-signup').value.trim();
-
-  if (email && password && parent_name && phone) {
-    const response = await fetch('/api/parents', {
-      method: 'post',
-      body: JSON.stringify({
-        parent_name,
-        phone,
-        email,
-        password
-      }),
-      headers: { 'Content-Type': 'application/json' }
-    });
-
-    if (response.ok && response_two.ok) {
-      document.location.replace('/parents/');
-    } else {
-      alert(response.statusText);
-    }
-  }
-}
-
 async function loginFormHandler(event) {
   event.preventDefault();
 
@@ -53,5 +25,5 @@ async function loginFormHandler(event) {
 }
 
 document.querySelector('.login-form').addEventListener('submit', loginFormHandler);
-document.querySelector('.signup-form').addEventListener('submit', signupFormHandler);
+
