+++ public/javascript/logout.js
@@ -1,5 +1,4 @@
 //Function to logout and return to main page/login
-
 async function logout() {
     const response = await fetch('/api/parents/logout', {
       method: 'post',
