+++ server.js
@@ -1,16 +1,21 @@
 const path = require('path');
 const express = require('express');
-const routes = require('./controllers');
 const sequelize = require('./config/connection');
+const routes = require('./controllers');
 
+//Handlbars
+const exphbs = require('express-handlebars');
+const hbs = exphbs.create({});
 
 const app = express();
 const PORT = process.env.PORT || 3001;
 
-
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 app.use(express.static(path.join(__dirname, 'public')));
+
+app.engine('handlebars', hbs.engine);
+app.set('view engine', 'handlebars');
 
 // turn on routes
 app.use(routes);
