+++ views/admin.handlebars
@@ -1,5 +1,5 @@
-// Admin View   -   wesbsite/admin
+{{!-- // Admin View   -   wesbsite/admin
 //admin-page-1.jpg
 //Upon login - the admin is directed here
 //View of all students 
-// all student info will be present ; include option to update status for student or delete
+// all student info will be present ; include option to update status for student or delete --}}
