+++ views/homepage.handlebars
@@ -1,4 +1,24 @@
-//Homepage  -     website 
-//refer to Main.jpg
-// This will contain a data input form that will create the parent model and associated child model
-// redirect parent to parent page after submit info is validated
+<form class="signup-form">
+
+    <div>
+        <label for="parent-name-signup">Parent Name:</label>
+        <input type="text" id="parent-name-signup" />
+    </div>
+    <div>
+        <label for="email-signup">Email:</label>
+        <input type="text" id="email-signup" />
+    </div>
+    <div>
+        <label for="password-signup">Password:</label>
+        <input type="password" id="password-signup" />
+    </div>
+    <div>
+        <label for="phone-signup">Phone Number:</label>
+        <input type="phone" id="phone-signup" />
+    </div>
+
+    <div>  
+    <button type="submit">Sign Up</button>
+    </div>
+
+</form>
