+++ views/layouts/main.handlebars
@@ -1,2 +1,37 @@
-//html design template that is used/visible on all pages (header/body/footer)
-//header changes based on who is logged in
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Jar-O-Kids</title>
+  <link rel="stylesheet" href="/stylesheets/style.css">
+</head>
+
+<body>
+  <div class="wrapper">
+    <header>
+      <h1>
+        <a href="/">Jar-O-Kids</a>
+      </h1>
+      <nav>
+        {{!-- {{#if loggedIn}}
+        <a href="/homepage">dashboard</a>
+        <button id="logout" class="btn-no-style">logout</button>
+        {{else}}
+        <a href="/login">login</a>
+        {{/if}} --}}
+      </nav>
+    </header>
+    <main>
+      {{{ body }}}
+    </main>
+    <footer>
+    </footer>
+  </div>
+  {{!-- {{#if loggedIn}}
+  <script src="/javascript/logout.js"></script>
+  {{/if}} --}}
+</body>
+
+</html>
