+++ views/parent.handlebars
@@ -1,5 +1,5 @@
-// Parent View   -   wesbsite/parent
+{{!-- // Parent View   -   wesbsite/parent
 //parent-login-page.jpg
 //View of all students associated with the parent
 // all student info will be present 
-//add student form
+//add student form --}}
